requirements.txt
alembic==1.12.0
blinker==1.6.3
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
Flask==3.0.0
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==23.2
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==3.0.0
